source/services/content/requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
starlette==0.20.4
uvloop==0.17.0
gevent==22.10.2
pydantic==1.10.2
aiohttp==3.8.3
alembic==1.8.1
anyio==3.6.2
ormar==0.11.3
asyncpg==0.26.0
databases==0.6.1
SQLAlchemy==1.4.41
attrs==22.1.0
bcrypt==4.0.1
click==8.1.3
ecdsa==0.18.0
exceptiongroup==1.0.4
fastapi-pagination==0.10.0
h11==0.14.0
httpcore==0.16.1
httpx==0.23.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
iniconfig==1.1.1
Mako==1.2.4
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.4
pyasn1==0.4.8
pyparsing==3.0.9
pytest==7.2.0
python-jose==3.3.0
python-multipart==0.0.5
pypinyin==0.47.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
typing_extensions==4.4.0
zipp==3.10.0
email-validator==1.3.0
PyYAML==6.0
esdk-obs-python==3.22.2
pycryptodome==3.10.1